# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jengo Building Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /buildings:
    get:
      operationId: listBuildings
      tags: [buildings]
      summary: List all buildings
      responses:
        "200":
          description: List of buildings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Building"
    post:
      operationId: createBuilding
      tags: [buildings]
      summary: Create a building
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBuildingBody"
      responses:
        "201":
          description: Created building
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{id}:
    get:
      operationId: getBuilding
      tags: [buildings]
      summary: Get a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Building
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
    patch:
      operationId: updateBuilding
      tags: [buildings]
      summary: Update a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBuildingBody"
      responses:
        "200":
          description: Updated building
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
  /buildings/{id}/stats:
    get:
      operationId: getBuildingStats
      tags: [buildings]
      summary: Get building stats (open issues, overdue payments, occupancy)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Building stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BuildingStats"

  # ── Units ──────────────────────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: buildingId
          in: query
          schema:
            type: integer
        - name: pinned
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementBody"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Visitors ───────────────────────────────────────────────────────────────
...
  /contractors:
    get:
      operationId: listContractors
      tags: [contractors]
      summary: List contractors
      responses:
        "200":
          description: List of contractors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contractor"
    post:
      operationId: createContractor
      tags: [contractors]
      summary: Create a contractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractorBody"
      responses:
        "201":
          description: Created contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
  /contractors/{id}:
    get:
      operationId: getContractor
      tags: [contractors]
      summary: Get contractor details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
...
components:
  schemas:
...
    Building:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        neighbourhood:
          type: string
          enum: [kilimani, westlands, south_b, lavington, parklands, karen, langata, other]
        totalUnits:
          type: integer
        managementCompany:
          type: string
        caretakerName:
          type: string
        caretakerPhone:
          type: string
        serviceChargeAmount:
          type: number
        reputationScore:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, address, neighbourhood, totalUnits, createdAt]
    CreateBuildingBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        neighbourhood:
          type: string
          enum: [kilimani, westlands, south_b, lavington, parklands, karen, langata, other]
        totalUnits:
          type: integer
        managementCompany:
          type: string
        caretakerName:
          type: string
        caretakerPhone:
          type: string
        serviceChargeAmount:
          type: number
      required: [name, address, neighbourhood, totalUnits]
    UpdateBuildingBody:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        caretakerName:
          type: string
        caretakerPhone:
          type: string
        serviceChargeAmount:
          type: number
    BuildingStats:
      type: object
      properties:
        buildingId:
          type: integer
        openIssues:
          type: integer
        inProgressIssues:
          type: integer
        overduePayments:
          type: integer
        occupancyRate:
          type: number
        avgResponseTimeHours:
          type: number
        activeContractors:
          type: integer
        pendingVisitors:
          type: integer
      required: [buildingId, openIssues, inProgressIssues, overduePayments, occupancyRate]
...
    Announcement:
      type: object
      properties:
        id:
          type: integer
        buildingId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [general, maintenance, utility, emergency, event, agm]
        isPinned:
          type: boolean
        authorName:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, buildingId, title, content, category, isPinned, createdAt]
    CreateAnnouncementBody:
      type: object
      properties:
        buildingId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [general, maintenance, utility, emergency, event, agm]
        isPinned:
          type: boolean
        authorName:
          type: string
        expiresAt:
          type: string
          format: date-time
      required: [buildingId, title, content, category]
    UpdateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        expiresAt:
          type: string
          format: date-time

    # ── Visitor ──
...
    Contractor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        trade:
          type: string
          enum: [plumbing, electrical, carpentry, painting, cleaning, security, landscaping, general, other]
        phone:
          type: string
        email:
          type: string
        rating:
          type: number
        totalJobs:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, trade, phone, createdAt]
    CreateContractorBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        trade:
          type: string
          enum: [plumbing, electrical, carpentry, painting, cleaning, security, landscaping, general, other]
        phone:
          type: string
        email:
          type: string
      required: [name, trade, phone]
...